Makefile update -Remove color, percent and all decoration -Rework make option SAN, NOERR, DEV -Fix auto dependency generator -Add Makefile as dependency

+++ Makefile
@@ -6,11 +6,10 @@
 #    By: chbravo- <chbravo-@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2015/12/08 11:02:51 by chbravo-          #+#    #+#              #
-#    Updated: 2017/10/30 23:15:57 by chbravo-         ###   ########.fr        #
+#    Updated: 2017/12/19 22:46:19 by chbravo-         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
-#TODO fix indent, space before '\'...
 NAME			= skell
 
 SRC_SUBDIR		= core
@@ -23,19 +22,18 @@ SRCS			+= main.c
 ###############################################################################
 #  Compiler
 CC				= clang
-CFLAGS			= -g -Wall -Wextra -Werror
+CFLAGS			= -Wall -Wextra
+
+ifneq ($(NOERR),yes)
+    CFLAGS		+= -Werror
+endif
 
 ifeq ($(DEV),yes)
-	CFLAGS		+= -std=c11 -pedantic -pedantic-errors
+    CFLAGS		+= -g
 endif
 
 ifeq ($(SAN),yes)
-	LDFLAGS 	+= -fsanitize=address
-	CFLAGS		+= -fsanitize=address -fno-omit-frame-pointer -fno-optimize-sibling-calls
-endif
-
-ifeq ($(NOERR),yes)
-    CFLAGG		= -g -Wall -Wextra -Wdeprecated-declarations
+    CFLAGS		+= -fsanitize=address -fno-omit-frame-pointer -fno-optimize-sibling-calls
 endif
 
 #The Directories, Source, Includes, Objects and Libraries
@@ -56,20 +54,7 @@ BUILD_DIR		= $(OBJS_DIR) $(DEPS_DIR)
 #Utils
 RM				= rm -rf
 MKDIR			= mkdir -p
-COUNT_OBJ		= 0
-COUNT_DEP		= 0
-TOTAL			= 0
-PERCENT			= 0
-$(eval TOTAL=$(shell echo $$(printf "%s" "$(SRCS)" | wc -w)))
 
-#color
-C_NO = \033[0m
-C_G = \033[0;32m
-C_Y = \033[1;33m
-C_B = \033[1;34m
-C_C = \033[1;36m
-C_R = \033[1;31m
-C_P = \033[1;35m
 DOXYGEN = $(shell doxygen -v dot 2> /dev/null)
 
 ###############################################################################
@@ -80,34 +65,28 @@ DOXYGEN = $(shell doxygen -v dot 2> /dev/null)
 #########
 ## RULES ##
  #########
-#.SECONDARY: $(OBJS) lib
-
-.NOTPARALLEL:
-all: $(DEPS) lib $(NAME)
+all: $(DEPS_DIR) lib $(NAME)
 
 # Add dependency as prerequisites
--include $(DEPS)
-
+ifneq ($(MAKECMDGOALS),clean)
+ ifneq ($(MAKECMDGOALS),fclean)
+  ifneq ($(MAKECMDGOALS),doc)
+   -include $(DEPS)
+  endif
+ endif
+endif
 
 $(NAME): $(OBJS)
-	@$(CC) $(CFLAGS) -o $(NAME) $(OBJS) $(LIBS) $(INC)
-	@echo "$(C_G)$(C_NO) ALL LINKED $(C_G)$(C_NO)"
+	$(CC) $(CFLAGS) -o $(NAME) $(OBJS) $(LIBS) $(INC)
 	@echo "INFO: Flags: $(CFLAGS)"
-	@echo "[\033[35m---------------------------------\033[0m]"
-	@echo "[\033[36m--------- SKELL Done ! ----------\033[0m]"
-	@echo "[\033[35m---------------------------------\033[0m]"
+	@echo "Done"
 
-$(OBJS_DIR)/%.o: %.c | $(OBJS_DIR)
-	@$(CC) $(LDFLAGS) $(CFLAGS) $(INC) -o $@ -c $<
-	$(eval COUNT_OBJ=$(shell echo $$(($(COUNT_OBJ)+1))))
-	$(eval PERCENT=$(shell echo $$((($(COUNT_OBJ) * 100 )/$(TOTAL)))))
-	@printf "$(C_B)%-8s $(C_P) $<$(C_NO)\n" "[$(PERCENT)%]"
+$(OBJS_DIR)/%.o: %.c Makefile | $(OBJS_DIR)
+	$(CC) $(LDFLAGS) $(CFLAGS) $(INC) -o $@ -c $<
 
+.PRECIOUS: $(DEPS_DIR)/%.d
 $(DEPS_DIR)/%.d: %.c | $(DEPS_DIR)
-	@$(CC) $(INC) -MM $< -MT $(OBJS_DIR)/$*.o -MF $@
-	$(eval COUNT_DEP=$(shell echo $$(($(COUNT_DEP)+1))))
-	$(eval PERCENT=$(shell echo $$((($(COUNT_DEP) * 100 )/$(TOTAL)))))
-	@printf "$(C_B)%-8s $(C_G) $@$(C_NO)\n" "[$(PERCENT)%]"
+		@$(CC) $(INC) -MM $< -MT $(OBJS_DIR)/$*.o -MF $@
 
 $(BUILD_DIR):
 	@$(MKDIR) -p $@
@@ -118,15 +97,17 @@ lib:
 re: fclean $(DEPS) lib $(NAME)
 
 clean:
-	@echo -e "\033[35m21sh  :\033[0m [\033[31mSuppression des .o\033[0m]"
-	@$(RM) $(OBJS_DIR)
-	@echo -e "\033[35m21sh  :\033[0m [\033[31mSuppression des .d\033[0m]"
-	@$(RM) $(DEPS_DIR)
+ifeq ($(shell [ -e $(OBJS_DIR) ] && echo 1 || echo 0),1)
+	$(RM) $(OBJS_DIR)
+endif
+ifeq ($(shell [ -e $(DEPS_DIR) ] && echo 1 || echo 0),1)
+	$(RM) $(DEPS_DIR)
+endif
 
 fclean: clean
-	@echo -e "\033[35m21sh  :\033[0m [\033[31mSuppression de $(NAME)\033[0m]"
-	@$(RM) $(NAME)
-	@rm -rf DOC
+ifeq ($(shell [ -e $(NAME) ] && echo 1 || echo 0),1)
+	$(RM) $(NAME)
+endif
 
 dev:
 	@make -C ./ SAN="yes" DEV="yes"
